requirements.txt
streamlit==1.30.0
lancedb==0.4.0
openai==1.6.1
python-dotenv==1.0.0
pypdf==4.0.0
langchain-text-splitters>=0.0.1
pyarrow==14.0.2
langchain==0.1.0
langchain_openai==0.0.2
langchain_community==0.0.13
uuid>=0.0.1
datetime
tiktoken==0.5.2
tavily-python==0.2.7
PyPDF2==3.0.1
langdetect==1.0.9
camelot-py[cv]==0.11.0
ghostscript==0.7
opencv-python==4.8.0.74
pytesseract==0.3.10
Pillow==10.0.0
docx2txt==0.8
pdfplumber==0.9.0
matplotlib==3.7.2
sympy==1.12
plotly>=5.18.0
python-docx==0.8.11
colorama>=0.4.6
PyMuPDF==1.22.5
python-dateutil>=2.8.2
scikit-learn>=1.2.2
pandas>=2.0.3
numpy>=1.24.3